Refactor: Rename some file
