+++ src/pages/AIgame/index.tsx
