+++ src/pages/OnlineGame/index.tsx
@@ -1,7 +1,9 @@
 import React, { useEffect } from 'react';
+import { useParams } from 'react-router-dom';
 import firestore from '../../utils/firestore';
 
 function OnlineGame() {
+  const roomID = useParams;
   useEffect(() => {
     // -
   });
